refactor: rename search album file, update paths
