+++ src/components/forms/album/index.js
@@ -1,6 +1,6 @@
 const buildAddAlbumForm = require('./addAlbum.js');
 const buildEditAlbumForm = require('./editAlbum.js');
-const searchAlbumForm = require('./editAlbumSearch.js');
+const searchAlbumForm = require('./searchAlbum.js');
 
 module.exports = {
   buildAddAlbumForm,
